backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.11.1
absl-py=2.1.0=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
bleach=6.2.0=py39haa95532_0
blinker=1.9.0=py39haa95532_0
bottleneck=1.4.2=py39hc99e966_0
brotli-python=1.0.9=py39h5da7b33_9
ca-certificates=2025.2.25=haa95532_0
certifi=2025.1.31=py39haa95532_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=py39haa95532_0
colorama=0.4.6=py39haa95532_0
comm=0.2.1=py39haa95532_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.11=py39h5da7b33_0
decorator=5.1.1=pyhd3eb1b0_0
exceptiongroup=1.2.0=py39haa95532_0
executing=0.8.3=pyhd3eb1b0_0
flask=3.1.0=py39haa95532_0
flask-cors=5.0.1=pypi_0
flatbuffers=25.2.10=pypi_0
fonttools=4.56.0=pypi_0
gast=0.6.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.70.0=pypi_0
h5py=3.13.0=pypi_0
idna=3.7=py39haa95532_0
importlib-metadata=8.5.0=py39haa95532_0
importlib-resources=6.5.2=pypi_0
importlib_metadata=8.5.0=hd3eb1b0_0
intel-openmp=2024.2.1=h57928b3_1083
ipykernel=6.29.5=py39haa95532_1
ipython=8.15.0=py39haa95532_0
itsdangerous=2.2.0=py39haa95532_0
jedi=0.19.2=py39haa95532_0
jinja2=3.1.6=py39haa95532_0
joblib=1.4.2=pyhd8ed1ab_1
jupyter_client=8.6.3=py39haa95532_0
jupyter_core=5.7.2=py39haa95532_0
kaggle=1.6.17=pyhd8ed1ab_1
kagglehub=0.2.7=py39haa95532_0
keras=3.8.0=pypi_0
kiwisolver=1.4.7=pypi_0
libblas=3.9.0=31_h641d27c_mkl
libcblas=3.9.0=31_h5e41251_mkl
libclang=18.1.1=pypi_0
libhwloc=2.11.2=default_ha69328c_1001
libiconv=1.18=h135ad9c_1
liblapack=3.9.0=31_h1aa476e_mkl
libsodium=1.0.18=h62dcd97_0
libwinpthread=12.0.0.r4.gg4f2fc60ca=h57928b3_9
libxml2=2.13.6=he286e8c_0
libzlib=1.3.1=h2466b09_2
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=py39h827c3e9_0
matplotlib=3.9.4=pypi_0
matplotlib-inline=0.1.6=py39haa95532_0
mdurl=0.1.2=pypi_0
mkl=2024.2.2=h66d3029_15
ml-dtypes=0.4.1=pypi_0
namex=0.0.8=pypi_0
nest-asyncio=1.6.0=py39haa95532_0
numexpr=2.10.2=mkl_py39h633a001_0
numpy=1.26.4=py39hddb5d58_0
openssl=3.4.1=ha4e3fda_0
opt-einsum=3.4.0=pypi_0
optree=0.14.0=pypi_0
packaging=24.2=py39haa95532_0
pandas=2.2.3=py39h5da7b33_0
parso=0.8.4=py39haa95532_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=11.1.0=pypi_0
pip=25.0=py39haa95532_0
platformdirs=3.10.0=py39haa95532_0
prompt-toolkit=3.0.43=py39haa95532_0
protobuf=5.29.3=pypi_0
psutil=5.9.0=py39h827c3e9_1
pure_eval=0.2.2=pyhd3eb1b0_0
pygments=2.15.1=py39haa95532_1
pyparsing=3.2.1=pypi_0
pysocks=1.7.1=py39haa95532_0
python=3.9.21=h8205438_1
python-dateutil=2.9.0post0=py39haa95532_2
python-slugify=5.0.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytz=2025.1=pypi_0
pywin32=308=py39h5da7b33_0
pyzmq=26.2.0=py39h5da7b33_0
requests=2.32.3=py39haa95532_1
rich=13.9.4=pypi_0
scikit-surprise=1.1.4=py39h941ed31_1
scipy=1.13.1=py39h1a10956_0
seaborn=0.13.2=pypi_0
setuptools=75.8.0=py39haa95532_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
tbb=2021.13.0=h62715c5_1
tensorboard=2.18.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.18.0=pypi_0
tensorflow-intel=2.18.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.5.0=pypi_0
text-unidecode=1.3=pyhd3eb1b0_0
tornado=6.4.2=py39h827c3e9_0
tqdm=4.67.1=py39h9909e9c_0
traitlets=5.14.3=py39haa95532_0
typing_extensions=4.12.2=py39haa95532_0
tzdata=2025.1=pypi_0
ucrt=10.0.22621.0=h57928b3_1
unidecode=1.3.8=py39haa95532_0
urllib3=2.3.0=py39haa95532_0
vc=14.42=haa95532_4
vc14_runtime=14.42.34433=h6356254_24
vs2015_runtime=14.42.34433=hfef2bbc_24
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39haa95532_1
werkzeug=3.1.3=py39haa95532_0
wheel=0.45.1=py39haa95532_0
win_inet_pton=1.1.0=py39haa95532_0
wrapt=1.17.2=pypi_0
zeromq=4.3.5=hd77b12b_0
zipp=3.21.0=py39haa95532_0
